requirements.txt
Faker==10.0.0
kafka-python==2.0.2
python-dateutil==2.8.2
six==1.16.0
text-unidecode==1.3